achieve fetch success condition of function

+++ src/com/unimelb/comp90015/fourLiterGroup/ezshare/serverOps/Resource.java
@@ -10,7 +10,7 @@ public class Resource {
 	private String Channel = null;
 	private String Owner = null;
 	private String[] EZserver = null;
-	private int resourceSize=0;
+	private long resourceSize=0;
 	public Resource() {// Constructor
 	}
 
@@ -42,7 +42,7 @@ public class Resource {
 		this.EZserver = ezservers.clone();
 	}
 	
-	public void setResourceSize(int size){
+	public void setResourceSize(long size){
 		this.resourceSize=size;
 	}
 	public String getName(){
@@ -73,7 +73,7 @@ public class Resource {
 		return this.Tags;
 	}
 	
-	public int getSize(){
+	public long getSize(){
 		return this.resourceSize;
 	}
 }
